feat: 优化日志记录和增加删除域名功能
- 增加 `tracing::error` 用于记录错误日志，并更新部分日志信息。
- 添加根据账号删除域名的功能 `delete_domain_by_account`。
- 更新依赖版本，包括 `tracing-test` 和其他相关库。
- 重构测试用例，使用 `tracing_test` 宏进行日志验证。

+++ crates/domain_manager/src/api/model/domain.rs
@@ -43,10 +43,10 @@ pub struct Domain {
     registration_date: String,
 
     #[serde(rename = "ExpirationCurrDateDiff")]
-    expiration_curr_date_diff: i32,
+    expiration_curr_date_diff: i64,
 
     #[serde(rename = "RegistrationDateLong")]
-    registration_date_long: i32,
+    registration_date_long: i64,
 
     #[serde(rename = "ResourceGroupId")]
     resource_group_id: String,
@@ -70,7 +70,7 @@ pub struct Domain {
     ccompany: String,
 
     #[serde(rename = "ExpirationDateLong")]
-    expiration_date_long: i32,
+    expiration_date_long: i64,
 
     #[serde(rename = "ExpirationDateStatus")]
     expiration_date_status: String,
